Extract common code to new trait
