+++ Magento2/Sniffs/Legacy/LayoutSniff.php
@@ -19,6 +19,8 @@ use SimpleXMLElement;
  */
 class LayoutSniff implements Sniff
 {
+    use ParseXMLTrait;
+
     private const ERROR_CODE_XML = 'WrongXML';
     private const ERROR_CODE_OBSOLETE_BLOCK = 'ObsoleteBlock';
     private const ERROR_CODE_OBSOLETE_CLASS = 'ObsoleteClass';
@@ -260,25 +262,6 @@ class LayoutSniff implements Sniff
                     );
                 }
             }
-        }
-    }
-
-    /**
-     * Format the incoming XML to avoid tags split into several lines.
-     *
-     * @param File $phpcsFile
-     *
-     * @return false|string
-     */
-    private function getFormattedXML(File $phpcsFile)
-    {
-        try {
-            $doc = new DomDocument('1.0');
-            $doc->formatOutput = true;
-            $doc->loadXML($phpcsFile->getTokensAsString(0, count($phpcsFile->getTokens())));
-            return $doc->saveXML();
-        } catch (\Exception $e) {
-            return false;
         }
     }
