+++ Magento2/Sniffs/Legacy/ObsoleteSystemConfigurationSniff.php
@@ -13,7 +13,8 @@ use PHP_CodeSniffer\Sniffs\Sniff;
 
 class ObsoleteSystemConfigurationSniff implements Sniff
 {
-    private const ERROR_CODE_XML = 'WrongXML';
+    use ParseXMLTrait;
+
     private const WARNING_CODE_OBSOLETE = 'FoundObsoleteSystemConfiguration';
 
     /**
@@ -54,43 +55,6 @@ class ObsoleteSystemConfigurationSniff implements Sniff
                 dom_import_simplexml($element)->getLineNo() - 1,
                 self::WARNING_CODE_OBSOLETE
             );
-        }
-    }
-
-    /**
-     * Adds an invalid XML error
-     *
-     * @param File $phpcsFile
-     * @param int $stackPtr
-     */
-    private function invalidXML(File $phpcsFile, int $stackPtr): void
-    {
-        $phpcsFile->addError(
-            "Couldn't parse contents of '%s', check that they are in valid XML format.",
-            $stackPtr,
-            self::ERROR_CODE_XML,
-            [
-                $phpcsFile->getFilename(),
-            ]
-        );
-    }
-
-    /**
-     * Format the incoming XML to avoid tags split into several lines.
-     *
-     * @param File $phpcsFile
-     *
-     * @return false|string
-     */
-    private function getFormattedXML(File $phpcsFile)
-    {
-        try {
-            $doc = new DomDocument('1.0');
-            $doc->formatOutput = true;
-            $doc->loadXML($phpcsFile->getTokensAsString(0, count($phpcsFile->getTokens())));
-            return $doc->saveXML();
-        } catch (\Exception $e) {
-            return false;
         }
     }
 }
