+++ Magento2/Sniffs/Legacy/WidgetXMLSniff.php
@@ -16,10 +16,11 @@ use PHP_CodeSniffer\Sniffs\Sniff;
  */
 class WidgetXMLSniff implements Sniff
 {
+    use ParseXMLTrait;
+
     private const ERROR_CODE_OBSOLETE_SUPPORTED_BLOCKS = 'FoundObsoleteNodeSupportedBlocks';
     private const ERROR_CODE_OBSOLETE_BLOCK_NAME = 'FoundObsoleteNodeBlockName';
     private const ERROR_CODE_FACTORY = 'FoundFactory';
-    private const ERROR_CODE_XML = 'WrongXML';
 
     /**
      * @inheritdoc
@@ -80,43 +81,6 @@ class WidgetXMLSniff implements Sniff
                 dom_import_simplexml($element)->getLineNo() - 1,
                 self::ERROR_CODE_OBSOLETE_BLOCK_NAME
             );
-        }
-    }
-
-    /**
-     * Adds an invalid XML error
-     *
-     * @param File $phpcsFile
-     * @param int $stackPtr
-     */
-    protected function invalidXML(File $phpcsFile, int $stackPtr): void
-    {
-        $phpcsFile->addError(
-            "Couldn't parse contents of '%s', check that they are in valid XML format",
-            $stackPtr,
-            self::ERROR_CODE_XML,
-            [
-                $phpcsFile->getFilename(),
-            ]
-        );
-    }
-
-    /**
-     * Format the incoming XML to avoid tags split into several lines.
-     *
-     * @param File $phpcsFile
-     *
-     * @return false|string
-     */
-    private function getFormattedXML(File $phpcsFile)
-    {
-        try {
-            $doc = new DomDocument('1.0');
-            $doc->formatOutput = true;
-            $doc->loadXML($phpcsFile->getTokensAsString(0, count($phpcsFile->getTokens())));
-            return $doc->saveXML();
-        } catch (\Exception $e) {
-            return false;
         }
     }
 }
